+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/contrib/views/PhoneNumberViewHolderFactoryInstrumentedTest.kt
@@ -42,7 +42,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class QuestionnaireItemPhoneNumberViewHolderFactoryInstrumentedTest {
+class PhoneNumberViewHolderFactoryInstrumentedTest {
   private lateinit var context: ContextThemeWrapper
   private lateinit var parent: FrameLayout
   private lateinit var viewHolder: QuestionnaireItemViewHolder
@@ -56,7 +56,7 @@ class QuestionnaireItemPhoneNumberViewHolderFactoryInstrumentedTest {
         R.style.Theme_Material3_DayNight
       )
     parent = FrameLayout(context)
-    viewHolder = QuestionnaireItemPhoneNumberViewHolderFactory.create(parent)
+    viewHolder = PhoneNumberViewHolderFactory.create(parent)
     questionnaireEditAdapter = QuestionnaireEditAdapter()
   }
