+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreEnumerationTest.kt
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.utilities
+package com.google.android.fhir.datacapture
 
+import com.google.android.fhir.datacapture.utilities.toCoding
 import com.google.common.truth.Truth.assertThat
 import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.Enumeration
