+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreLocalDateTimesTest.kt
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.utilities
+package com.google.android.fhir.datacapture
 
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
+import com.google.android.fhir.datacapture.utilities.format
+import com.google.android.fhir.datacapture.utilities.toLocalizedTimeString
 import com.google.common.truth.Truth.assertThat
 import java.time.LocalDate
 import java.time.LocalDateTime
