+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreReflectionUtilTest.kt
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.utilities
+package com.google.android.fhir.datacapture
 
+import com.google.android.fhir.datacapture.utilities.invokeFunction
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
